requirements.txt
Django==3.2.5
djangorestframework==3.12.4
pandas==1.3.1
openpyxl>=3.0.0,<4.0.0
